barcode-windows/requirements.txt
hidapi==0.14.0
pystray==0.19.5
Pillow==10.4.0

